+++ crates/sclang-lsp/src/features/mod.rs
@@ -2,6 +2,7 @@ pub mod completion;
 pub mod diagnostics;
 pub mod document_highlight;
 pub mod find_references;
+pub mod folding_range;
 pub mod goto;
 pub mod hover;
 pub mod inlay_hints;